requirements.txt
chardet==5.2.0
python-debian==0.1.49
python-dotenv==1.0.1
